+++ src/components/MiniPlayer.jsx
@@ -1,30 +1,12 @@
-import { useRef } from "react";
-
 import CloseIcon from "@/assets/svgs/icon-close.svg?react";
 import PauseIcon from "@/assets/svgs/icon-mini-pause.svg?react";
 import PlayIcon from "@/assets/svgs/icon-mini-player.svg?react";
-import { SETTING_TYPES } from "@/constants/settingOptions";
-import { useChannelStore } from "@/store/useChannelStore";
-import { useUserStore } from "@/store/useUserStore";
-import controlStreamingPlayback from "@/utils/playControl";
+import useChannelPlayback from "@/hooks/useChannelPlayback";
 
 const MiniPlayer = ({ closePlayer }) => {
-  const { selectedChannelId, radioChannelList, isPlaying, setIsPlaying } =
-    useChannelStore();
-  const { settings } = useUserStore();
-  const isAdDetect = settings[SETTING_TYPES.AD_DETECT];
-
-  const selectedChannel = radioChannelList.find(
-    ({ id }) => id === selectedChannelId
-  );
+  const { videoId, selectedChannel, isPlaying, handlePlayPause } =
+    useChannelPlayback();
   const { name, logoUrl } = selectedChannel;
-
-  const videoId = useRef(null);
-
-  const handlePlayPause = () => {
-    controlStreamingPlayback(videoId, selectedChannelId, isPlaying, isAdDetect);
-    setIsPlaying((prev) => !prev);
-  };
 
   return (
     <div className="flex h-20 w-full justify-between rounded-b-4xl bg-white px-4 shadow-[0_-6px_9px_rgba(0,0,0,0.3)]">
